refactor(chat): split model selector and add mobile drawer
Extract state, keyboard handling, and UI into dedicated modules. On viewports
below the md breakpoint, use a bottom sheet (Vaul drawer) instead of a popover
so the picker is easier to use on small screens while keeping the popover on
desktop.
